fileglob: move checking for ~ and / into path_dwim_get_candidates
Instead of checking for paths relative to the home directory or
absolute paths twice (in path_dwim_relative_stack and in the fileglob
plugin), move the whole check into path_dwim_get_candidates and use
only this function in the fileglob plugin.

This change only looks this huge in dataloader.py because the
indentation changed and the  identical comments make git lose the
correct context for the code. Actually, this only moves the check
for ~ and / around and changes the parameter because b_source is not
needed in path_dwim_relative_stack anymore.